readjusted the image.js
